requirements.txt
blinker==1.6.3
build==1.0.3
certifi==2024.6.2
click==8.1.7
elastic-transport==8.13.1
elasticsearch==8.14.0
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
packaging==23.2
pip-tools==7.3.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
urllib3==2.2.2
Werkzeug==3.0.1
